Added fixes

+++ planner/optaplanner-shiftplanning/src/main/java/com/kairos/shiftplanning/domain/activity/Activity.java
@@ -44,7 +44,7 @@ public class Activity {
     private String name;
     @JsonIgnore
     private Map<ConstraintSubType, ConstraintHandler> constraints;
-    private Set<Long> validDayTypeIds;
+    private Set<BigInteger> validDayTypeIds;
     private TimeType timeType;
     private int order;
     private int activityPrioritySequence;
